Move `checkedToString` to the reflection module
... since it's reflection functionality. Also allows for deduplicating
code with testing.js
