+++ testing.js
@@ -3,7 +3,7 @@
  * @license MPL-2.0
  */
 
-import { isStringable, toArrayIfNecessary } from "./src/reflection.js";
+import { checkedToString, toArrayIfNecessary } from "./src/reflection.js";
 
 /**
  * A list of example shell injection strings to test whether or not a function
@@ -27,13 +27,7 @@ export const injectionStrings = ["\x00world", "&& ls", "'; ls #", '"; ls #'];
  * - Converts non-array inputs to single-item arrays where necessary.
  */
 export const shescape = {
-  escape: (arg, _options) => {
-    if (!isStringable(arg)) {
-      throw new TypeError();
-    } else {
-      return arg.toString();
-    }
-  },
+  escape: (arg, _options) => checkedToString(arg),
   escapeAll: (args, _options) => {
     args = toArrayIfNecessary(args);
     return args.map(shescape.escape);
